+++ src/main/java/tbooop/model/boss/impl/meaty/Meaty.java
@@ -1,4 +1,4 @@
-package tbooop.model.boss.impl.Meaty;
+package tbooop.model.boss.impl.meaty;
 
 import tbooop.commons.impl.HealthImpl;
 import tbooop.model.boss.api.AbstractBoss;
@@ -8,7 +8,7 @@ import tbooop.model.player.api.Player;
 public class Meaty extends AbstractBoss<MeatySM.State> {
 
     public Meaty(Player p) {
-        super(new HealthImpl(30), 0, 14);
+        super(new HealthImpl(45), 0, 14);
         setStateMachine(new MeatySM(this, p));
     }