+++ src/main/java/tbooop/model/boss/impl/meaty/MeatySM.java
@@ -1,4 +1,4 @@
-package tbooop.model.boss.impl.Meaty;
+package tbooop.model.boss.impl.meaty;
 
 import tbooop.model.boss.stateMachine.api.AbstractStateMachine;
 import tbooop.model.player.api.Player;
@@ -7,12 +7,14 @@ import java.util.Map;
 public class MeatySM extends AbstractStateMachine<MeatySM.State> {
 
     public enum State {
-        BASE;
+        SPIRAL,
+        BURST;
     }
 
     public MeatySM(Meaty meaty, Player p) {
-        super(State.BASE);
-        setStates(Map.of(State.BASE, new MeatyBase(meaty)));
+        super(State.SPIRAL);
+        setStates(Map.of(State.BURST, new MeatyBurst(meaty, p),
+                State.SPIRAL, new MeatySpiral(meaty, p)));
         init();
     }
 }